feat(frontend): Add 404 error page
Introduce a custom 404 error page (frontend\error.vue) to enhance user experience in case of navigation to non-existing routes. The page includes a friendly message and a link to the homepage for easy redirection. The styling is optimized for both desktop and small screens to ensure a consistent and visually pleasing layout.